requirements.txt
# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Processing and Manipulation
numpy==1.24.3
pandas==2.0.3
pillow==10.0.0
opencv-python==4.8.0.74
scikit-image==0.21.0
h5py==3.9.0

# Machine Learning and Statistics
scikit-learn==1.3.0
scipy==1.11.1
statsmodels==0.14.0

# GAN Evaluation Metrics
pytorch-fid==0.3.0
lpips==0.1.4
clean-fid==0.1.35

# Visualization and Plotting
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Experiment Tracking and Logging
tensorboard==2.13.0
wandb==0.15.8

# Progress Bars and Utilities
tqdm==4.65.0
pyyaml==6.0.1

# Jupyter and Interactive Development
jupyter==1.0.0
ipykernel==6.25.0
notebook==7.0.2

# Testing and Quality
pytest==7.4.0
black==23.7.0
flake8==6.1.0

# Additional Utilities
kaggle==1.5.16
gdown==4.7.1
